Add DeferralAllowed fields to PlanOpts, graphWalkOpts, and Deferred
The concept behind deferred actions is that various objects (resources, data
sources, modules, stack components) can end up being *deferred*, meaning that
we're unable to produce an actionable plan for them and intend to run the *rest*
of the plan to completion and apply it without taking action on those objects,
which will have to wait for another future plan.

For the most part, we only *find out* about the kinds of absences that force
deferral at the time of *graph walk.* That's because deferral is caused by
unknown values in count/for_each, absent dependencies reported by the provider,
or dependency on something that got deferred, none of which can really be know
at graph construction.

Also, we don't always want deferral to be an available option -- for the time
being, we want to preserve existing behavior in the modules runtime ("normal
terraform") while enabling deferral in Stacks.

So, *things that happen during graph walk need to know whether they can defer.*
(And if not, they'll throw an error if they hit such an absence.) That means the
top-level signal that deferral is enabled (which should arrive as part of the
PlanOpts) needs to make it all the way down into the graph walk methods.

The graph walk methods generally have access to an EvalContext, and EvalContext
conveniently gives you access to a Deferred tracker, the pointer to which is
shared across all related EvalContexts. So, the Deferred will be what tells you
whether you can defer.